task-driver: tasks: node-startup: Define node startup task skeleton
