+++ common/Cargo.toml
@@ -47,6 +47,9 @@ constants = { path = "../constants" }
 renegade-crypto = { path = "../renegade-crypto" }
 util = { path = "../util" }
 
+# === Contracts Repo Dependencies === #
+contracts-common = { git = "https://github.com/renegade-fi/renegade-contracts.git" }
+
 # === Misc Dependencies === #
 base64 = { version = "0.13" }
 bimap = "0.6.2"
