+++ workers/task-driver/src/tasks/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod create_new_wallet;
 pub mod lookup_wallet;
+pub mod node_startup;
 pub mod pay_offline_fee;
 pub mod pay_relayer_fee;
 pub mod redeem_relayer_fee;
